implementada consulta ao clima mostrando na tela
